% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{DTconvert2factor}
\alias{DTconvert2factor}
\title{Bulk type conversion of columns to factors}
\usage{
DTconvert2factor(x, typeToConvert = "chr")
}
\arguments{
\item{x}{data.frame (or data.table)}

\item{typeToConvert}{type of the columns to be converted}
}
\value{
data.table
}
\description{
Converts multiple columns of same type to factor
Possible enhancements : Specify the columns to be converted
}
\examples{
DT <- data.frame(c1=c("a","b","c")
,c11 = c("a","b","c")
,c2 = 1:3
,c3 = c(1.1,1.2,1.3)
,stringsAsFactors = F)
DT <- DTconvert2factor(DT,"int")
}
\seealso{

}
\author{
Amit Agni
}
\concept{type conversion functions}
