% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{fiscal_quarter}
\alias{fiscal_quarter}
\title{Assign fiscal quarter to a date}
\usage{
fiscal_quarter(x, fiscal_start = 1)
}
\arguments{
\item{x}{Date for which the fiscal quarter is desired}

\item{fiscal_start}{The month in which the fiscal year begins e.g 7 for Australia}
}
\value{
numeric vector with the fiscal year and the quarter
}
\description{
Returns the fiscal quarter for the given date. This function is a modified version of lubridate::quarter.
The original function had an issue and the solution was copied from \url{https://github.com/tidyverse/lubridate/issues/682}
}
\examples{
fiscal_quarter(Sys.Date())
}
\seealso{
\url{https://github.com/tidyverse/lubridate/issues/682}  for the original issue
}
\concept{data time functions}
