% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OS.R
\name{killProcess}
\alias{killProcess}
\title{Kills or ends a windows process}
\usage{
killProcess(process_name)
}
\arguments{
\item{x}{process to kill}
}
\value{
None
}
\description{
The task name can be retrieved from the Windows Task list.
Useful for killing EXCEL processes that linger-on even after Excel application is quit
}
\examples{
killProcess("EXCEL")
}
\seealso{


Other operating system functions: 
\code{\link{write2clip}()}
}
\author{
Amit Agni
}
\concept{operating system functions}
