% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{outlookSend}
\alias{outlookSend}
\title{Send email from outlook}
\usage{
outlookSend(
  To,
  Subject = "",
  CC = NULL,
  BCC = NULL,
  HTMLBodyFile = NULL,
  HTMLBody = NULL,
  Body = "",
  Attachments = NULL,
  DeferredDeliveryTime = NULL,
  OriginatorDeliveryReportRequested = NULL,
  ReadReceiptRequested = NULL,
  FlagRequest = NULL,
  Importance = NULL,
  Sensitivity = NULL,
  SentOnBehalfOfName = NULL
)
}
\arguments{
\item{To}{Receiver}
}
\description{
Uses VBA script to send emails from Outlook. Attach files or insert Rmarkdown generated html files in the body of the email
All credit goes to the solution from  \url{https://stackoverflow.com/questions/26811679/sending-email-in-r-via-outlook}
}
\examples{
outlookSend(To=c("xyz@gmail.com","abc@gmail.com")
,Subject="Daily Ticketed Intakes (Revenue)"
,HTMLBodyFile="filename.html")
}
\seealso{
\url{https://stackoverflow.com/questions/26811679/sending-email-in-r-via-outlook}  for the original post
}
\concept{email functions}
