% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{runSQLinFile}
\alias{runSQLinFile}
\title{Runs the query given in a text file on  postgreSQL database}
\usage{
runSQLinFile(
  sqlFilePath,
  readFromRds,
  bkpExisting = T,
  user = NA,
  password = NA,
  dbname,
  host,
  port
)
}
\arguments{
\item{sqlFilePath}{Location of the Sql file}

\item{readFromRds}{Looks for the .Rds file in the same path as sqlFilePath}

\item{bkpExisting}{Backup the existing RDS file}

\item{user}{database user name, if not provided will be prompted}

\item{password}{database passowrd, if not provided will be prompted}

\item{dbname}{database name}

\item{host}{database host name}

\item{port}{database port}
}
\value{
data.table
}
\description{
Connects to postgresSQL database and runs the query given in the specified file.
Also, stores the results in a .RDS file in the same location as the sqlFilePath
Possible enhancements : Option to not store as RDS
}
\examples{
runPostgreSQL_inFile(sqlFileName = here("SQL_get-dates.sql"),user = "",password="",readFromRds = F,dbname = "dw",host = "dwserver", port = 5050)
}
\seealso{
\code{\link{runSQLinText}}

Other database functions: 
\code{\link{runSQLinText}()}
}
\author{
Amit Agni
}
\concept{database functions}
