% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{runSQLinText}
\alias{runSQLinText}
\title{Runs the query given as sqlText string}
\usage{
runSQLinText(sqlText, user = NA, password = NA, dbname, host, port)
}
\arguments{
\item{sqlText}{Sql query}

\item{user}{database user name, if not provided will be prompted}

\item{password}{database passowrd, if not provided will be prompted}

\item{dbname}{database name}

\item{host}{database host name}

\item{port}{database port}
}
\description{
Connects to postgresSQL database and runs the query given in the string.
Possible enhancements : Store as RDS
}
\examples{
runPostgreSQL_asText(sqlText = "select * from table limit 19",user = "",password="",readFromRds = F,dbname = "dw",host = "dwserver", port = 5050)
}
\seealso{
\code{\link{runPostgreSQL_inFile}}

Other database functions: 
\code{\link{runSQLinFile}()}
}
\author{
Amit Agni
}
\concept{database functions}
