% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{theme_darklightmix}
\alias{theme_darklightmix}
\title{a ggplot theme with a mixture of dark and light backgrounds}
\usage{
theme_darklightmix(
  font_size = 10,
  legend_position = "right",
  color_theme = "light"
)
}
\arguments{
\item{font_size}{Base font size}

\item{legend_position}{the position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{color_theme}{Choose from a combination of light and dark themes ("va_light","va_dark","light","lightcyan","cornsilk","antiquewhite","darkslategray","gray25","turquoise4","steelblue")}
}
\description{
A theme based on ggplot theme_minimal. Uses default R colors for plot background
}
\examples{
for(i in c("va_light","va_dark"
,"light","lightcyan","cornsilk","antiquewhite"
,"darkslategray","gray25","turquoise4","steelblue")){
print(
 iris \%>\%
    ggplot(aes(x=Sepal.Length,y=Sepal.Width,color=factor(paste("Petal",round(Petal.Length,0))))) +
      geom_point() +
      labs(x="Sepal Length", y="Sepal Width",
        title="Petals and Sepals",ro
          subtitle="Plot shows the sizes of petals and sepals",
          caption=i,
          color = "Legend") +
          facet_wrap(~Species,scales = "free") +
          theme_darklightmix(color_theme = i) +
          scale_color_brewer(type="qual",palette = "Set3")  )
          }
}
\concept{ggplot functions}
