% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OS.R
\name{write2clip}
\alias{write2clip}
\title{Write to clipboard}
\usage{
write2clip(x, row.names = FALSE, col.names = TRUE, ...)
}
\arguments{
\item{x}{object to copy}
}
\value{
None
}
\description{
Not tested on Mac
}
\examples{
vec <-"test"
write2clip(vec)

}
\seealso{


Other operating system functions: 
\code{\link{killProcess}()}
}
\concept{operating system functions}
